backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==1.0.1
psycopg2-binary==2.9.9
google-generativeai==0.3.0